Dev test

+++ src/com/sgaop/project/module/NutzBootMakerChooseStep.java
@@ -1,82 +1,68 @@
 package com.sgaop.project.module;
 
 import com.google.gson.Gson;
-import com.intellij.ide.util.projectWizard.ProjectJdkStep;
+import com.intellij.ide.util.projectWizard.ModuleWizardStep;
 import com.intellij.ide.util.projectWizard.WizardContext;
 import com.intellij.ide.wizard.CommitStepException;
 import com.intellij.openapi.options.ConfigurationException;
-import com.sgaop.project.FileUtil;
+import com.intellij.openapi.ui.Messages;
+import com.sgaop.project.HttpUtil;
 import com.sgaop.project.ui.ModuleWizardStepUI;
-import org.apache.commons.io.IOUtils;
+import com.sgaop.util.FileUtil;
 import org.apache.http.HttpResponse;
-import org.apache.http.client.HttpClient;
-import org.apache.http.client.methods.HttpGet;
-import org.apache.http.client.methods.HttpPost;
 import org.apache.http.entity.BasicHttpEntity;
-import org.apache.http.impl.client.HttpClients;
 
 import javax.swing.*;
 import java.io.ByteArrayInputStream;
 import java.io.File;
-import java.io.IOException;
 import java.nio.file.Paths;
 import java.util.HashMap;
 
 /**
  * @author 黄川 huchuc@vip.qq.com
  */
-public class NutzBootMakerChooseStep extends ProjectJdkStep {
+public class NutzBootMakerChooseStep extends ModuleWizardStep {
 
     protected final WizardContext wizardContext;
 
-    private ModuleWizardStepUI moduleWizardStepUI;
+    private ModuleWizardStepUI moduleWizardStepUI = new ModuleWizardStepUI();
 
     private String downLoadKey;
+
+    private boolean loadingCompleted = false;
 
     private Gson gson = new Gson();
 
     public NutzBootMakerChooseStep(WizardContext wizardContext) {
-        super(wizardContext);
         this.wizardContext = wizardContext;
-        this.moduleWizardStepUI = new ModuleWizardStepUI();
-    }
-
-    @Override
-    public JComponent getComponent() {
-        return moduleWizardStepUI.getRoot();
     }
 
     @Override
     public void onWizardFinished() throws CommitStepException {
-        HttpClient httpclient = HttpClients.createDefault();
-        HttpGet httppost = new HttpGet(moduleWizardStepUI.getMakerUrl().getText() + "/maker/download/" + downLoadKey);
-        try {
-            String path = this.wizardContext.getProjectFileDirectory();
-            if (this.wizardContext.getProject() != null) {
-                NutzBootModuleBuilder moduleBuilder = (NutzBootModuleBuilder) this.wizardContext.getProjectBuilder();
-                path = moduleBuilder.getPath();
+        if (isNutzBootModuleBuilder()) {
+            try {
+                HttpResponse response = HttpUtil.getResponse(moduleWizardStepUI.getMakerUrl().getText() + "/maker/download/" + downLoadKey);
+                String path = this.wizardContext.getProjectFileDirectory();
+                if (this.wizardContext.getProject() != null) {
+                    path = ((NutzBootModuleBuilder) this.wizardContext.getProjectBuilder()).getContentEntryPath();
+                }
+                File zipFile = Paths.get(path, "temp.zip").toFile();
+                File dir = Paths.get(path).toFile();
+                FileUtil.writeFile(zipFile, response.getEntity().getContent());
+                FileUtil.extractZipFile(zipFile, dir);
+                zipFile.delete();
+            } catch (Exception e) {
+                throw new CommitStepException(e.getMessage());
             }
-            HttpResponse response = httpclient.execute(httppost);
-            File zipFile = Paths.get(path, "temp.zip").toFile();
-            File dir = Paths.get(path).toFile();
-            FileUtil.writeFile(zipFile, response.getEntity().getContent());
-            FileUtil.extractZipFile(zipFile, dir);
-            zipFile.delete();
-        } catch (IOException e) {
-            throw new CommitStepException(e.getMessage());
         }
     }
 
     @Override
     public boolean validate() throws ConfigurationException {
-        HttpClient httpclient = HttpClients.createDefault();
-        HttpPost httppost = new HttpPost(moduleWizardStepUI.getMakerUrl().getText() + "/maker/make");
         try {
             BasicHttpEntity entity = new BasicHttpEntity();
             entity.setContent(new ByteArrayInputStream(gson.toJson(moduleWizardStepUI.getPostData()).getBytes()));
-            httppost.setEntity(entity);
-            HttpResponse response = httpclient.execute(httppost);
-            String json = IOUtils.toString(response.getEntity().getContent(), "UTF-8");
+            String json = HttpUtil.post(moduleWizardStepUI.getMakerUrl().getText() + "/maker/make", entity);
             HashMap redata = gson.fromJson(json, HashMap.class);
             boolean ok = Boolean.parseBoolean(String.valueOf(redata.getOrDefault("ok", "false")));
             if (ok) {
@@ -84,11 +70,25 @@ public class NutzBootMakerChooseStep extends ProjectJdkStep {
                 return true;
             } else {
                 downLoadKey = null;
-                JOptionPane.showMessageDialog(null, redata.getOrDefault("msg", "服务暂不可用！请稍后再试！"), "错误提示", JOptionPane.ERROR_MESSAGE, null);
+                Messages.showErrorDialog(redata.getOrDefault("msg", "服务暂不可用！请稍后再试！").toString(), "错误提示");
                 return false;
             }
         } catch (Exception e) {
             throw new ConfigurationException("构筑中心发生未知异常! " + e.getMessage());
+        }
+    }
+
+
+    @Override
+    public void updateStep() {
+        if (!loadingCompleted) {
+            try {
+                String json = HttpUtil.get(moduleWizardStepUI.getMakerUrl().getText() + "/maker.json");
+                moduleWizardStepUI.refresh(json);
+                loadingCompleted = true;
+            } catch (Exception e) {
+                Messages.showErrorDialog("网络异常！请稍候尝试!" + e.getMessage(), "错误提示");
+            }
         }
     }
 
@@ -97,28 +97,11 @@ public class NutzBootMakerChooseStep extends ProjectJdkStep {
     }
 
     @Override
-    public void onStepLeaving() {
+    public JComponent getComponent() {
+        return moduleWizardStepUI.getRoot();
     }
 
-    @Override
-    public boolean isStepVisible() {
-        return true;
-    }
-
-    @Override
-    public void disposeUIResources() {
-    }
-
-    @Override
-    public void updateStep() {
-        HttpClient httpclient = HttpClients.createDefault();
-        HttpGet httppost = new HttpGet(moduleWizardStepUI.getMakerUrl().getText() + "/maker.json");
-        try {
-            HttpResponse response = httpclient.execute(httppost);
-            String json = IOUtils.toString(response.getEntity().getContent(), "UTF-8");
-            moduleWizardStepUI.refresh(json);
-        } catch (Exception e) {
-            JOptionPane.showMessageDialog(null, "网络异常！请稍候尝试", "错误提示", JOptionPane.ERROR_MESSAGE, null);
-        }
+    private boolean isNutzBootModuleBuilder() {
+        return this.wizardContext.getProjectBuilder() instanceof NutzBootModuleBuilder;
     }
 }
